+++ index.php
@@ -1,6 +1,16 @@
 <?php
+
 // load dependencies
-require_once(__DIR__ . '/vendor/autoload.php');
+if(is_file($f = __DIR__ . '/vendor/autoload.php')) {
+    // local composer install
+    require_once($f);
+} elseif(is_file($f = __DIR__ . '/../../../vendor/autoload.php')) {
+    // root composer install
+    require_once($f);
+} else {
+    // composer needs install...
+    die('Cannot find composer `/vendor/autoload.php` -- try `composer install`');
+}
 
 // instance Pico
 $pico = new Pico(